Fix bug when specifying boolean values -- fixes issue #7
We were only accepting literal 0 or 1 as a boolean value, but of course
Perl has a richer syntax for this, so we now accept other boolean values
as well. Thanks to Alex Pankoff (ccntrq in github) for reporting.

While fixing this, I realized the tests were actually faulty (sob), so
fixed that and added tests for the new boolean value support.