Allow parsing cookies with names that have no value (fixes issue #5)
The parser was rejecting cookies where some fields have no value.  This
is of course wrong, particularly for fields such as HttpOnly.

Reported by Nigel Gregoire (nfg in github); thanks!

I fixed this by remembering whether the parser saw an equals sign or not
for each field, and returning that fact; if we got a field with no
equals sign, then we associate an undef with it.  This allows us to
support fields with an empty value, by specifying the equals sign.

Added tests for all these cases, which were failing before the patch but
now work fine.