Support multi-value cookies (fixes issue #3)
We now accept a cookie with a value that is an arrayref.  The values in
the array get joined with '&', and then the whole string is URL-encoded.

When crushing such a cookie, the reverse applies.

This was suggested by James Raspass (JRaspass in github); thanks!

The implementation is a bit dirty, I need to review it.