Rework Buffer implementation
Buffers used to have a single internal position.  This made it difficult
to use them when one wants to first write into the buffer and then read
back for further processing.

I changed this to use two positions: rpos and wpos, and reimplemented
everything to take advantage of this.  The code is now simpler, and it
can be made even simpler, which is my next step.

I also removed the macros in Buffer that tried to make them coexist with
C strings, they were error prone.