Optionally accept names without values
The standard (RFC 6265) seems to say that when crushing a cokie from a
string, it is not allowed to specify names without a value (so you would
not be allowed to include stuff like HttpOnly and Secure).

Since there has been some discussion (for example, in Dancer2) as to
what is the correct approach, I have added an optional parameter to
crush_cookie that allows the caller to say "please give me names without
a value, and set them to undef".  The default value is zero, so you will
not get these valueless names.